+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added
+- Client statements for Scroll, MouseOver, and Window Resize events. [#243](https://github.com/zaproxy/zest/issues/243)
 
 ## [0.17.0] - 2023-03-31
 ### Changed